syllabus: added link to cleipnir slides, removed bitcoin, reorg
This change adds the link to the cleipnir guest lecture slides
that were already added to the slides folder. Also removes the
reference to the bitcoin lecture that have been taken out of
syllabus since the topic is covered in the security course.

This also reorganizes the optional material in separate section at
the bottom of the list.